feat: move payload types into crate::payload
