+++ src/payload/mod.rs
@@ -1,0 +1,33 @@
+//! Contains request payload datatypes.
+
+mod agent;
+mod catalog;
+
+use std::time::Duration;
+
+pub use agent::*;
+pub use catalog::*;
+
+#[derive(Clone, Debug, Default)]
+pub struct WriteOptions {
+    pub datacenter: Option<String>,
+}
+
+
+#[derive(Clone, Debug, Default)]
+pub struct QueryOptions {
+    pub datacenter: Option<String>,
+    pub wait_index: Option<u64>,
+    pub wait_time: Option<Duration>,
+}
+
+#[derive(Clone, Debug)]
+pub struct QueryMeta {
+    pub last_index: Option<u64>,
+    pub request_time: Duration,
+}
+
+#[derive(Clone, Debug)]
+pub struct WriteMeta {
+    pub request_time: Duration,
+}
